+++ devbin/transpile.sh
@@ -17,5 +17,17 @@
 #      '.| /      <https://github.com/v0lkan/sif/issues>.
 #
 
-# `source` this for development purposes.
-alias s=./bin/sif.js
+PD=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
+
+cd $PD/..
+
+FILENAME=$( basename "$1" )
+EXTENSION="${FILENAME##*.}"
+FILENAME="${FILENAME%.*}"
+DIR=$( dirname "$1" )
+
+echo "babel --source-maps --out-file $DIR/$FILENAME.js $1"
+
+babel --source-maps --out-file "$DIR/$FILENAME.js" $1
+
+echo "Done."